+++ geomanager/helpers.py
@@ -1,74 +1,6 @@
 from django.contrib.admin.utils import quote
 from django.urls import reverse
-from django.utils.translation import gettext as _
-from wagtail_modeladmin.helpers import ButtonHelper, AdminURLHelper
-
-
-class CategoryButtonHelper(ButtonHelper):
-    def get_buttons_for_obj(
-            self, obj, exclude=None, classnames_add=None, classnames_exclude=None
-    ):
-        buttons = super().get_buttons_for_obj(obj, exclude, classnames_add, classnames_exclude)
-
-        classnames = self.edit_button_classnames + classnames_add
-        cn = self.finalise_classname(classnames, classnames_exclude)
-
-        create_dataset_button = {
-            "url": obj.dataset_create_url(),
-            "label": _("Create Dataset"),
-            "classname": cn,
-            "title": _("Create Dataset") % {"object": self.verbose_name},
-        }
-
-        buttons.append(create_dataset_button)
-
-        return buttons
-
-
-class DatasetButtonHelper(ButtonHelper):
-    def get_buttons_for_obj(
-            self, obj, exclude=None, classnames_add=None, classnames_exclude=None
-    ):
-        buttons = super().get_buttons_for_obj(obj, exclude, classnames_add, classnames_exclude)
-
-        classnames = self.edit_button_classnames + classnames_add
-        cn = self.finalise_classname(classnames, classnames_exclude)
-
-        layer_create_url = obj.create_layer_url()
-
-        if layer_create_url:
-            create_layer_button = {
-                "url": obj.create_layer_url(),
-                "label": _("Add Layer"),
-                "classname": cn,
-                "title": _("Add %(object)s Layer") % {"object": self.verbose_name},
-            }
-            buttons.append(create_layer_button)
-
-        return buttons
-
-
-class FileLayerButtonHelper(ButtonHelper):
-    def get_buttons_for_obj(
-            self, obj, exclude=None, classnames_add=None, classnames_exclude=None
-    ):
-        buttons = super().get_buttons_for_obj(obj, exclude, classnames_add, classnames_exclude)
-
-        classnames = self.edit_button_classnames + classnames_add
-        cn = self.finalise_classname(classnames, classnames_exclude)
-
-        url = obj.get_style_url()
-
-        layer_style_button = {
-            "url": url.get("url"),
-            "label": _(url.get("action")),
-            "classname": cn,
-            "title": _(url.get("action")) % {"object": self.verbose_name},
-        }
-
-        buttons.append(layer_style_button)
-
-        return buttons
+from wagtail_modeladmin.helpers import AdminURLHelper
 
 
 def get_layer_action_url(layer_type, action, action_args=None):
