+++ geomanager/templates/geomanager/modeladmin/index_without_custom_create.html
@@ -1,5 +1,5 @@
 {% extends "wagtailadmin/base.html" %}
-{% load i18n modeladmin_tags wagtailadmin_tags %}
+{% load i18n modeladmin_tags wagtailadmin_tags geomanager_tags %}
 
 {% block bodyclass %}modeladmin{% endblock %}
 
@@ -17,7 +17,11 @@
 
 {% block content %}
     {% block header %}
-        {% fragment as breadcrumb %}{% block breadcrumb %}{% endblock %}{% endfragment %}
+        {% fragment as breadcrumb %}{% block breadcrumb %}
+            {% if navigation_items %}
+                {% breadcrumbs_component navigation_items %}
+            {% endif %}
+        {% endblock %}{% endfragment %}
         {% fragment as heading %}{% block h1 %}{% endblock %}{% endfragment %}
         {% fragment as description %}{% include 'modeladmin/includes/result_count.html' %}{% endfragment %}
         {% fragment as extra_actions %}
