+++ geomanager/templates/geomanager/raster_file/raster_file_edit_form.html
