+++ geomanager/templates/geomanager/stations/stations_preview.html
