+++ geomanager/templates/geomanager/stations/stations_upload.html
