+++ geomanager/templates/geomanager/vector_file/vector_file_edit_form.html
