+++ geomanager/templates/geomanager/vector_file/vector_file_upload.html
